Add more docs and move some fns to display
