+++ instanseg/inference_class.py
@@ -80,6 +80,67 @@ def _display_colourized(mIF):
     return colour_render.astype(np.uint8)
 
 
+
+def download_model(model_str: str, 
+                   verbose: bool = True, 
+                   headers: Optional[str] = None):
+    """
+    Download a model from the specified GitHub release.
+
+    :param model_str: The model string to download.
+    :param verbose: Whether to print verbose output.
+    :param headers: Optional headers for the request, eg for authentication.
+    """
+    import os
+    import requests
+    import zipfile
+    from io import BytesIO
+    import torch
+
+
+    if not os.environ.get("INSTANSEG_BIOIMAGEIO_PATH"):
+        os.environ["INSTANSEG_BIOIMAGEIO_PATH"] = os.path.join(os.path.dirname(__file__),"../bioimageio_models/")
+
+    bioimageio_path = os.environ.get("INSTANSEG_BIOIMAGEIO_PATH")
+
+
+    # Ensure the directory exists
+    os.makedirs(bioimageio_path, exist_ok=True)
+    
+    release_tag = "instanseg_models_v1"
+    url = f"https://api.github.com/repos/instanseg/instanseg/releases/tags/{release_tag}"
+    response = requests.get(url, headers=headers)
+    response.raise_for_status()  # Raise an error for bad response
+
+    release_data = response.json()
+    assets = release_data.get("assets", [])
+
+    if model_str in [asset["name"].replace(".zip","") for asset in assets if asset["name"].endswith(".zip")]:
+        url = r"https://github.com/instanseg/instanseg/releases/download/instanseg_models_v1/{}.zip".format(model_str)
+        response = requests.get(url)
+        response.raise_for_status()  # Raise an error for bad responses
+        with zipfile.ZipFile(BytesIO(response.content)) as z:
+            z.extractall(bioimageio_path)
+
+
+        if verbose:
+            print(f"Model {model_str} downloaded and extracted to {bioimageio_path}")
+
+        path_to_torchscript_model = bioimageio_path + f"{model_str}/instanseg.pt"
+        return torch.jit.load(path_to_torchscript_model)
+
+    else:
+
+        #load model locally
+
+        path_to_torchscript_model = os.path.join(bioimageio_path, model_str, "instanseg.pt")
+
+        if os.path.exists(path_to_torchscript_model):
+            return torch.jit.load(path_to_torchscript_model)
+        else:
+            raise Exception(f"Model {model_str} not found in the release data or locally. Please check the model name and try again.")
+
+
 class InstanSeg():
     """
     Main class for running InstanSeg.
